+++ src/main/backends/linux/x11/backend.ts
@@ -26,11 +26,12 @@ export class X11Backend implements Backend {
   /**
    * Override this if another type is more suitable for your desktop environment.
    * https://www.electronjs.org/docs/latest/api/browser-window#new-browserwindowoptions
-   *
-   * @returns 'dock'
    */
-  public getWindowType() {
-    return 'dock';
+  public getBackendInfo() {
+    return {
+      windowType: 'dock',
+      supportsShortcuts: true,
+    };
   }
 
   /** This is called when the backend is created. Currently, this this does nothing on X11. */
@@ -105,11 +106,11 @@ export class X11Backend implements Backend {
    * This binds a shortcut. The action callback of the shortcut is called when the
    * shortcut is pressed. On X11, this uses Electron's globalShortcut module.
    *
-   * @param shortcut The shortcut to simulate.
+   * @param shortcut The shortcut to bind.
    * @returns A promise which resolves when the shortcut has been bound.
    */
   public async bindShortcut(shortcut: Shortcut) {
-    if (!globalShortcut.register(shortcut.accelerator, shortcut.action)) {
+    if (!globalShortcut.register(shortcut.trigger, shortcut.action)) {
       throw new Error('Invalid shortcut or it is already in use.');
     }
   }
@@ -120,7 +121,7 @@ export class X11Backend implements Backend {
    * @param shortcut The shortcut to unbind.
    */
   public async unbindShortcut(shortcut: Shortcut) {
-    globalShortcut.unregister(shortcut.accelerator);
+    globalShortcut.unregister(shortcut.trigger);
   }
 
   /** This unbinds all previously bound shortcuts. */